Hooz 331 choices (#745)
* First phase HOOZ-331-Choices

first phase for shakedown testing

* Part 2 implementation

Part 2 implementation - all user defined choises

* Tranche 3

All requirements done - needs rigerous testing

* tranche 4

rubocop and best practices

* invalid relation

removed plot_choices

* Fix Rails Best Practices Routing report

* trailing whitespace..

* Get tests running

* Choice Tests added

Choices Tests

* missing last newline

* factories

* Changed orientation and layout

* Styling

* Review comments

* Added comments

* resolve merge conflict change

* Stop reference to null object while downloading CSV

+++ app/controllers/room_configurations/room_choice_controller.rb
@@ -103,8 +103,8 @@ module RoomConfigurations
       send_data(csv_data,
                 type: "plain/text",
                 filename: "Plot #{plot.number}-#{Time.zone.today.strftime('%d %B %Y')}.csv")
-      ensure
-        temp_file.unlink
+    ensure
+        temp_file.unlink if temp_file.present?
     end
     # rubocop:enable all
